toolkit/components/passwordmgr/test/mochitest/test_bug_776171.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=776171
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 776171 related to HTTP auth</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="startTest()">
<script class="testbody" type="text/javascript">

/**
 * This test checks we correctly ignore authentication entry
 * for a subpath and use creds from the URL when provided when XHR
 * is used with filled user name and password.
 *
 * 1. connect auth2/authenticate.sjs that expects user1:pass1 password
 * 2. connect a dummy URL at the same path
 * 3. connect authenticate.sjs that again expects user1:pass1 password
 *    in this case, however, we have an entry without an identity
 *    for this path (that is a parent for auth2 path in the first step)
 */

SimpleTest.waitForExplicitFinish();

function doxhr(URL, user, pass, next) {
  var xhr = new XMLHttpRequest();
  if (user && pass) {
    xhr.open("POST", URL, true, user, pass);
  } else {
    xhr.open("POST", URL, true);
  }
  xhr.onload = function() {
    is(xhr.status, 200, "Got status 200");
    next();
  };
  xhr.onerror = function() {
    ok(false, "request passed");
    SimpleTest.finish();
  };
  xhr.send();
}

function startTest() {
  doxhr("auth2/authenticate.sjs?user=user1&pass=pass1&realm=realm1", "user1", "pass1", function() {
    doxhr("auth2", null, null, function() {
      doxhr("authenticate.sjs?user=user1&pass=pass1&realm=realm1", "user1", "pass1", SimpleTest.finish);
    });
  });
}
</script>
</body>
</html>
